time:int->long
Former-commit-id: 77089adfb88319d0f651ddc6caf9ab23d3c93c34

+++ back-end-projects/Explorer/src/main/java/com/github/ontio/service/IBlockService.java
@@ -67,5 +67,5 @@ public interface IBlockService {
      * @param time
      * @return
      */
-    Result blockCountInTwoWeeks(int time);
+    Result blockCountInTwoWeeks(long time);
 }
